:recycle: refactor: 각 Entity와 Repository 별 Member Entity Mapping
